Move sudo-functionality to sudo.py
