+++ python/nav/django/context_processors.py
@@ -24,7 +24,8 @@ from operator import attrgetter
 from django.conf import settings
 
 from nav.config import find_config_file
-from nav.web.auth import get_sudoer, get_login_url, get_logout_url
+from nav.web.auth import get_login_url, get_logout_url
+from nav.web.auth.sudo import get_sudoer
 from nav.django.utils import get_account, is_admin
 from nav.web.message import Messages
 from nav.web.webfront.utils import tool_list, quick_read, split_tools
