+++ python/nav/web/auth/utils.py
@@ -1,0 +1,30 @@
+# Copyright (C) 2010, 2011, 2013, 2019 Uninett AS
+# Copyright (C) 2022 Sikt
+#
+# This file is part of Network Administration Visualized (NAV).
+#
+# NAV is free software: you can redistribute it and/or modify it under
+# the terms of the GNU General Public License version 3 as published by
+# the Free Software Foundation.
+#
+# This program is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+# FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
+# more details.  You should have received a copy of the GNU General Public
+# License along with NAV. If not, see <http://www.gnu.org/licenses/>.
+#
+
+import logging
+
+
+_logger = logging.getLogger(__name__)
+
+
+ACCOUNT_ID_VAR = 'account_id'
+
+
+def _set_account(request, account):
+    request.session[ACCOUNT_ID_VAR] = account.id
+    request.account = account
+    _logger.debug('Set active account to "%s"', account.login)
+    request.session.save()
